Chatbot Python Project
We are going to build an exciting project on Chatbot. We will implement a chatbot from scratch that will be able to understand what the user is talking about and give an appropriate response.

Prerequisites
To implement the chatbot, we will be using Keras, which is a Deep Learning library, NLTK, which is a Natural Language Processing toolkit, and some helpful libraries. Run the below command to make sure all the libraries are installed:
                                         pip install tensorflow keras pickle nltk

Project File Structure
1.Train_chatbot.py
2.Gui_Chatbot.py
3.Intents.json
4.Chatbot_model.h5
5.Classes.pkl
6.Words.pkl